feat: 视频画面单例模式
